requirements.txt
sqlalchemy==2.0.43
pandas==2.3.0
scikit-learn==1.7.0
python-dotenv==1.1.1
unidecode==1.4.0
numpy==2.3.1
dotenv==0.9.9
pyyaml==6.0.2